2013_11.10 - Updates / Long message to John
John - I think my issues are:
1. Since I can not read code (yet), I do not understand it.  And since I
don't understand it, I can't write it.
2. I am, therefore, stuck copying and pasting examples from the jQuery
API and / or following examples in the readings - and then trying them
out by trial and error.
3.  I also don't understand the structure of the program.  I looked at
Keith Lamar's Hot or Cold since you liked it.  (a) I then put the
variables and functions at the top and the
$(document).ready(function(){}); after it.  (b) After working all
morning on this, I still don't understand it.  What does
$(document).ready(function(){}); actually do - and why are all the
variables, functions, and if statements outside it?
4. I assume "scoping" means in lay terms routines and sub-routines.  My
read is that the random variable is "top-level" / parent and the guess
variable is child / inside the program.  That way if someone resets /
starts over, the random variable is retained and only the guesses start
over.  Correct?  Or am I missing this entirely.  Incidentally, I looked
at all the other students' work and I don't see any examples of nested
functions as shown in the readings (except Keith's, I think) - with the
caveat, that I have major difficulty understanding what I'm reading!